The AI is finally here
After the long awaiting of it the AI for the chatbot is finally here. It is not exactly the smartest one out there but it sort of gets the job done. In the next few days there are going to be commits regarding that because as of now it has to train every time it gets answered to. The goal is for it to learn before it initiates a conversation.