Very small changes
Increased the batch size and epochs so the bot's learning will be more efficient. Note that due to technical circumstances on my part the bot cannot be at it's full potential since the dataset that is being used is huge but in reality for the learning part there are only 450.